fix ScanRadius

+++ common/src/main/java/com/github/litermc/vsmecha/block/radar/RadarBlockEntity.java
@@ -135,7 +135,7 @@ public abstract class RadarBlockEntity extends EnergyBasedBlockEntity {
 	@Override
 	public void load(final CompoundTag data) {
 		super.load(data);
-		this.radius = data.getDouble("ScanRadius");
+		this.radius = data.getInt("ScanRadius");
 		this.scanRemaning = data.getInt("ScanRemaning");
 		this.autoScan = data.getBoolean("AutoScan");
 		this.queuingScan = data.getBoolean("QueuedScan");
@@ -145,7 +145,7 @@ public abstract class RadarBlockEntity extends EnergyBasedBlockEntity {
 	@Override
 	protected void saveAdditional(final CompoundTag data) {
 		super.saveAdditional(data);
-		data.putDouble("ScanRadius", this.radius);
+		data.putInt("ScanRadius", this.radius);
 		data.putInt("ScanRemaning", this.scanRemaning);
 		data.putBoolean("AutoScan", this.autoScan);
 		data.putBoolean("QueuedScan", this.queuingScan);
